+++ frontend/src/components/FlightItem.jsx
@@ -23,11 +23,13 @@ function FlightItem() {
     }
 
     const removeFlight = (e, index) => {
-        const data = {
-            category: e.target.id,
-            index
+        if(window.confirm('Are you sure you want to delete?')){
+            const data = {
+                category: e.target.id,
+                index
+            }
+            dispatch(removeCategoryItem(data))
         }
-        dispatch(removeCategoryItem(data))
     }
 
     return (
