+++ frontend/src/components/LodgingItem.jsx
@@ -24,12 +24,14 @@ function LodgingItem() {
     }
 
     const removeHotel = (e, index) => {
-        const data = {
-            category: e.target.id,
-            index
+        if(window.confirm('Are you sure you want to delete?')){
+            const data = {
+                category: e.target.id,
+                index
+            }
+    
+            dispatch(removeCategoryItem(data))
         }
-
-        dispatch(removeCategoryItem(data))
     }
