+++ frontend/src/components/NoteItem.jsx
@@ -19,11 +19,13 @@ function NoteItem() {
     }
 
     const removeNote = (e, index) => {
-        const data = {
-            category: e.target.id,
-            index
+        if(window.confirm('Are you sure you want to delete?')){
+            const data = {
+                category: e.target.id,
+                index
+            }
+            dispatch(removeCategoryItem(data))
         }
-        dispatch(removeCategoryItem(data))
     }   
 
     return (
